Add auth-skipping for more realistic debugging
I needed a way to be genuinely logged in as a user. Some application behavior honors an email parameter for authenticated users—showing their assignemntss—, but I wanted to debug something that was occurring at the moment of login, when the cookie is generated and an email is used to find a contact id. This helps make it easier to debug auth issues and remain close to what happens in reality.