1094: Fix bug on data feedback
There was a bug in our data feedback submission modal where a user's
feedback was not being posted to the backend. This bug was caused
by: when the new version of the zap-api was created, the feedback
route was not transferred over to the new version from the old version.

This commit re-adds this feedback route in project.controller.ts.

Octokit is used in the project.service.ts in a function called
sendFeedbackToGithubIssue in order to submit issues to our github
repository, dcp-zap-data-feedback.

This function is then run in the project.controller.ts in a @Post.
Because we use recaptcha in our frontend data feedback modal,
(frontend component `modal-controls.hbs`) we check that recapatcha has
not errored in order to run the sendFeedbackToGithubIssue function.