1097 Set up a Disposition creation endpoint for internal use
The endpoint is to support creating a Disposition to which
files will be uploaded. Files that are potentially being migrated
from DigitalOcean should be uploaded to their corresponding
Disposition. However, most Projects in CRM do not yet have a
coresponding Disposition for the incoming files, so we need to first create
the Disposotions via the CRM Web API during a scripted upload of files to
CRM.